Refactor for 0.3; standardized naming convention; specs done.
Pretty major overhaul that just standardizes how context behaves and what methods are called to invoke that behavior. Also allows for the ability to include mixins for specific contexts. Documentation updated to match.

Should make development with this library more predictable and easier to follow. Also provides the ability to build fluent APIs on top of contexts.