Repurpose `for` to set context name. Refactor `chain` into `use`.
Previously `for` did one too many things including: setting a context, binding that context to a function and then returning the chained result. Now it does considerably less and it is more clean what it does: it now just provides an object to chain middleware _for_ a certain context (or contexts). Code that previously used `ctx.for(mw1)` now has to rewrite it as `ctx.for('name').use(mw1)`. Other than that the behavior is just about the same. It gains the ability to accept obscure combinations of arguments for multiple contexts, so `ctx.for('a', ['b', 'c'], 'd')` is fine and flattened automatically.

The `chain` function has also received minor rework, now behaving much more like the express `use` call, so it has been renamed to match. `use` accepts middleware in the same manner as express does (flattened and run in serial), except that these middleware are ran once for each context previously set by `for`. So `ctx.for('a', 'b').use(mw1, mw2)` will run `mw1 -> mw2` in series bound to the context `'a'` _and_ `mw1 -> mw2` in series bound to the context `'b'`.

Overall this simplifies and clarifies use of the API and reduces the amount of code in the library.